Move the clamp function to the test where it's being used
