--- tests/cover/test_sizes.py
@@ -1,27 +1,0 @@
-# coding=utf-8
-#
-# This file is part of Hypothesis, which may be found at
-# https://github.com/HypothesisWorks/hypothesis-python
-#
-# Most of this work is copyright (C) 2013-2015 David R. MacIver
-# (david@drmaciver.com), but it contains contributions by others. See
-# CONTRIBUTING.rst for a full list of people who may hold copyright, and
-# consult the git log if you need to determine who owns an individual
-# contribution.
-#
-# This Source Code Form is subject to the terms of the Mozilla Public License,
-# v. 2.0. If a copy of the MPL was not distributed with this file, You can
-# obtain one at http://mozilla.org/MPL/2.0/.
-#
-# END HEADER
-
-from __future__ import division, print_function, absolute_import
-
-from hypothesis.utils.size import clamp
-
-
-def test_clamp():
-    assert clamp(None, 1, None) == 1
-    assert clamp(None, 10, 1) == 1
-    assert clamp(1, 0, 1) == 1
-    assert clamp(1, 0, None) == 1
